feat: clawe.workspaces ns, clawe.awesome/awm-fn impl
awm-fn supports converting clojure datastructures to lua syntax. This
makes it really convenient to pass data from clojure to a lua function.

clawe.workspaces gets a handler that will support updating the
workspaces widget. This handler supports calling a passed function name
with the latest workspace data, which is just hard-coded for now.