feat: building clawe as an uberscript
Creates a bin/clawe that executes the /clawe-script.clj with whatever
arguments. This is expected to be symlinked to your path somewhere.

Updates clawe.core to support building and installing the binary, so
that rebuilds are a viable clawe command. This should be refactored into
a useful tool somewhere, as it's the same logic/style that ralphie and
vapor also use.

Once installed, clawe can now be rebuilt via `clawe rebuild-clawe`.