misc: helpers added to util namespace
try-catch wrapper and log_if_error helpers are a nice win for error
handling.

these util functions are underscored to be sure they can be called from
lua ergonomically. kind of annoying, wonder if there's a way to smooth
that out, maybe by exporting both snake and camel-cased function names.
Or a macro for fennel modules that just turns a file of functions into a
proper namespace.