feat: create first client after creating workspace
A bit tentative about this, but i think it makes sense to auto-create a
client in every case where the workspace is empty (no clients) and it is
opened. This is roughly the same behavior as the scratchpad toggling,
and will reveal cases where workspaces don't have the right initial-file
or exec set.