wip: more defworkspace refactoring
Now supporting accepting xorfs, which is named b/c it's either x (some
map of data) or a f (a unary function x->x).

One thing we'll likely need is an ability to push some of these
functions to be evaluated at run-time, right now these are evaluated at
macro time, which will not be desired at some point.
If the partitioning that is easy, we should have some state of
workspaces that can be recalculated cheaply/measurably.