feat: impl toggle-all-titlebars
This was delightful. Learned that because I've already created and built
the defbinding, I don't need to re-write clawe-bindings.fnl or reload
awesome to update the effect of the same keybinding.

Feels like writing out to clawe-bindings.fnl could happen on save - it'd
be great if the 'restart' of awesomewm to rebuild the bindings was more
transparent. more reason to find something else to manage the keybindings.