feat: defkbds with quoted-fennel write pure awm kbds
When a defkbd's final xorf is a call to `(awm/awm-fnl
"some-fennel-string"),
the string is printed as the keybinding in
awesome/clawe-bindings.fnl. Normally kbds print as spawn-fn calls that
execute clawe itself, but now we're priting pure awesome functions as
well. This makes a performance difference for commands that could be to
the metal (or at least skipping the overhead of running clawe's
defcom/run command).