feat: toggle scratchpad - apply rules if client found elsewhere
Often a client ends up on the wrong tag, and the current fix is a
nuclear cmd-r, which happens to reapply the :rules/apply fns. This adds
a bit of logic to the toggle-scratchpad function to apply the function
for that workspace if the client is found in another workspace. This
happens most often for firefoxdeveloperedition, which ends up on the web
tag b/c the 'firefox' rule grabs it.