feat: include unmatched tags in all-workspaces
Tags are created on the fly in a few cases, as a fallback to allow
clients to be usable. One method is to launch a terminal when not
attached to any workspace (move to a scratchpad and then toggle it off).

This updates the clawe.workspaces/all-workspaces function to create
pseudo-workspaces for tags that didn't match a passed workspace. The
result is seeing the temp tags in topbar, and supporting deleting those
extra tags in clean-up commands.

These pseudo-workspaces are pretty handy - probably a good starting
place for on-the-fly workspace creation. Perhaps the doctor workspaces
widget should support changing these workspace names and setting other
details, like a workspace directory/repo, initial file, todo sources,
etc. All that requires reading/merging workspaces from the clawe.db.