feat: workspace db crud and :display-name support
Now supporting updating workspace names on the fly. Click in the topbar
workspace cell to edit in-place - this will create a workspace entry in
the datalevin db on-blur. db workspaces are merged into existing
workspaces or on-the-fly awm tags.

Still plenty of issues - the on-the-fly awm workspaces move around b/c
their names change when their index changes.... need a better way to mix
this data.