fix: create missing sxhkd session if it doesn't exist
Not perfect - i'd hoped to avoid this to keep keybindings closer to the
metal, speed-wise. It could probably be sped up, but this isn't the slow
part for now, and 'fast' keybindings can be done via awm/fnl anyway.

The main win is keybindings should now work after login, without
requiring a `clawe reload`. imagine that!