feat: support inlined sxhkd commands
Similar to the inlined-awm bindings, these are sxhkd bindings that get
executed directly (rather than run the whole clawe uberjar just to do
some shell command).

Moves a few bindings to using `bb -x` to fire the commands, which is
slightly quicker, execution wise. Note that this style uses the live
version of the code, whereas the uberjar is somewhat 'stable' or
'outdated', depending on your point of view. This means bindings can
fail if clawe is currently being refactored and in a broken state, which
might be annoying. It also means 'fixes' work right away, which is a
nice feedback loop. Maybe eventually we'll reach for some way to toggle
between these types of interactions - the stable and edge versions.