refactor: moves the workspace logic into wm itself
Rather than use `clawe.workspace` as an extra ns layer, this moves the
`clawe.workspace` helpers and public funcs into `clawe.wm`, and moves to
consuming `clawe.wm` everywhere.

The `clawe.wm` functions were really easy to reach for but as impled
were a weird extra layer. This is much nicer to consume.

This also introduces tests for the `wm` layer itself - these are kind of
fun b/c they run with whatever the current window-manager is, so it'll
simple to use them to debug/automate testing for whatever silly things
we hit.

Includes some malli test helpers.