fixes: wm-test passing in yabai!
Handful of little issues and missing yabai claweWM impl.

Includes a wait-for helper that re-runs an expression until it returns
truthy, which helps deal with yabai returning before osx is finished
switching workspaces or moving clients around.

The integration tests can take a bit on osx, something like 10 seconds
or so.