feat: toggle term/emacs working again, and data-driven
A bit complicated - here we extend client/match to support a
use-workspace-title style match, so that clients can be toggled per
workspace (if they opt-in). This match requires passing a workspace
title into client/match - i.e. what is the in-context workspace? A
question raise is what workspace to pass in wm/merge-client-defs - do
clients know their workspace name already? if they did, it might be
simple to use at that client/match call-site. For now we merge it in
toggle/find-client.