fix: rofi select shortest match
ralphie/rofi has had this bug forever! When rofi labels start with the
same prefix, it would select the first starts-with? match, which could
end up as a different item than the selected. Especially common when
selecting a workspace that starts with the same label (e.g. clerk and
clerk-demo).