feat: consume clerk html wrapper, send current path
Cleans up the log-state clerk status helper, which helps debugging quite
a bit.

When a client connects, it now sends a lone {:path doc.pathname} after
from it's 'onopen' - this lets us manage which channel is connected to
which notebook, so that we can update the currently watching clients
when that notebook is re-evaled.

Should probably just use a simple ch->notebook map as the datastructure,
and group-by notebook as needed - that'd clear up the hairy 'update'
logic.