fix: notebooks working at startup, better fallback route handling
Adds a basic list of links for `/`, and an error page when notebook eval
crashes.

Also restores notebooks/notebooks. This was working after being evaled
in the repl - it turns out requiring it sets *file* to
notebooks/core.clj - i was using it for the file prefix, but it was
crashing in fs/as-file without it. Will have to find a better way to do
that - probably by collecting the namespaces, but we don't necessarily
have them loaded yet :/