wip: adds multiple cljs testing approaches
Tests on the jvm via kaocha should be fine.

For cljs, kaocha has two cljs libs to consider - included is code for
the first one, which is not working, nor is it shadow-cljs compatible.

shadow-cljs provides a test build and runner of sorts via karma, which
is working here, and olical/cljs-test-runner worked with the least
amount of config (tho it might need more work to run in a browser).

these tests are not yet passing - i'm capturing notes on these
approaches to be shared. ideally they can all run via kaocha, which
looks like it requires another side process... but that's not too
different from karma.

it'd also be a boon to run integration tests via etaoin - in the future.