fix: handle str-ifying unixPaths before transit
Transit errors in plasma are a real pain to debug - i wonder if there's
a tap/portal/logging integration that could make it easier to figure out
what's going wrong when this blows up in the socket on-message.

This could be expensive (we're checking every outgoing val to see if
it's an instance of Path :/) - a few (fs/home)s were wrapped in strings,
which is probably enough to fix the issue on its own. A UnixPath transit
read/write handler is probably the better way. Really i just wish
transit would nil pun unsupported types.