focus: show the tag label for the 'love' tag group
A few considerations for a tag group component layout. Feels like i
want a list of components more than a single div, but maybe it's fine to
wrap elems arbitrarily. mostly i worry about css width inheritance, and
would prefer to not introduce a heriarchy just for a set of
components... but maybe it should be multiple components, not just one.