feat: show subtodos as cards in current-stack items
Also fixes when condition in item-body (with actions list) that was
hiding based on :org/body-string. This now also shows the item-body if
there are any :org/items.

Includes a sketch of a :presets idea for the filter-grouper, which
should lead to some immediate gains, e.g. a 'today' preset that restores
an approximation of the previous focus widget behavior (showing todos
just from the current daily).