feat: add :match-str-includes-any, fix default preset
Extends the filter-grouper's match api with a new match type,
~:match-str-includes-any~. the match api supports a :match-fn predicate
already, which is flexible, but not completely data (b/c functions show
up as Object[blah] when you stringify them in cljs). Supporting them as
data (edn) as much as possible is a win and keeps things really simpler.

Also fixes the default preset usage, which should really only be a
fallback for when local storage didn't find a last-set preset for a
given filter-grouper-id.