feat: org watcher purges edited/removed items
Supports basic sync-and-purge for a file. When an org item is edited,
the watcher syncs the latest to the database, and retracts db items
that belonged to the edited file but are no longer found in the latest
parsed version of that file.

We may need to push these retractions to the frontend if we aren't
already.