feat: remove deleted tags when purging org files
Extends garden.db's sync-and-purge to remove deleted tags. Similar
functionality needs to be applied to all lists.

An alternative is diffing the parsed org file with the current db to
figure out what to add/update/remove, then transacting that difference.
I was kind of hoping datascript could do that for me, but perhaps not.