feat: include parent relations for all org items
Not just those with uuids/parent-ids. This is done via other-db-updates,
which can be extended to include other note datoms that don't work as a
map in the note. (like datoms belonging to other entities, e.g.
children).

Also updates the filter-grouper to pass the active-filter-fn and passed
filter-items fn into the item->comp as :filter-by and :filter-fn. Not in
love with the naming, but i need these funcs passed to apply to children
when rendering todo/card-or-card-group.