feat: adding (and dissoc-ing) inferred actions on todos
Adds an initial inferred action for the views/todos widget: quick add
for tags on other items in the group.

Introduces an :action/popup-comp that pops up a div when you click the
action.

Unfortunately :actions/inferred cannot be passed into defhandlers ->
they crash because transit chokes on the :action/on-click function, or
some other nonsense that it can't deal with.