refactor: drop 'current' usage everywhere
Moves away from using a 'current' tag for :status/in-progress. It's
redundant, and shouldn't be necessary.

Leaving some things in place for a bit, in case there's a reason to
revert - I think at one point I liked supporting both so that 'current'
could be used on non-todo items, but at this point todos and notes have
a strong line between them (no longer mixed).