perf: cache mx-commands, read cache first in mx-fast
Getting kind of crazy, but closer to a very performant mx command,
via caching. Here we write the mx-command labels to a cache, and read
them instead of building them up when invoking clawe-mx. This lets us
have a slim-dependency mx_fast namespace that will load very quickly,
and once a command is selected, we create a new process for selecting
the command and then firing the 'on-select' for it.

This doesn't yet handle workspace-specific or context-specific
commands (like those based on current browser tabs). I have a hope we
can handle those asynchronously, at least on the rofi side.

This is still a bit slow - still takes a full second for rofi to render
175 items. digging there next.