feat: fast clawe-mx
Finally gets getting some nice performance out of clawe-mx.

This breaks rofi's MRU sorting b/c the rofi input string build is now
memoized. The performance win is worth it for now.

Moves 'dynamic' parts of clawe-mx to a mx-suggestions command on a
different keybinding. These can't be memoized (b/c their inputs change)
and break the performance gain, so a separate keybinding resolves some
of that.

The suggestions could have a server-supported version as well at some
point, which should be faster out of the box.

Cache-busting and cache-warming follow ups can come later on.