feat: mod+x/w for server-mx, mod+shift+x/w for bb-mx
Some fixes and clean up to the mx commands, which are now running well.

Adds a clawe-mx call to open-new-workspace, which warms up the cache
when the workspace is first opened. This might also be a good time to
clear the cache, and maybe 'restart-workspace' ends up becoming a
cache-clearing tactic.