feat: clawe-toggle via server, with fallback for journal
Moves clawe-toggle on awm to use the api and server, which is quicker
already, but could impl some workspace/client caching to be even faster.

Note that a journal-bb, terminal-bb fallback is left in place so we can
still toggle an emacs/terminal instance before the server starts
up (which will make it possible to debugging if the server is
down/crashing).