wip: import i3 config
Prepping to symlink and edit this directly from clawe.

Kind of funny to move this into clawe after finally getting the
clawe.edn OUT of here (and back into dotfiles). But, i think it makes
sense if the awesome and yabai configs are in here - really should be
working out a way to get those expressed in the clawe.edn and exported
into whatever the contents of these dirs are, and maybe gitignoring
them. Alternatively, clawe could just make it easier to write those, and
the dirs could provide a base config for other folks to extend on - e.g.
norms for fast-impls of clawe toggle or mx.