feat: start a new pomodoro every time we clawe-reload
I don't reload often, so this might be a nice boon. I suspect we'll have
to be careful about the sometimes-end-pomodoro datas, but i think going
loosey goosey on the data will free up more opportunities than holding
back on the data at this point.

Really this should be a smart-pomo-toggle helper that can get called
whenever. One step closer to automagic-pomodoros and 'modes' (in this
case, break-vs-working ui modes).