feat: add async git fetch via tmux helper
A helper for firing git-fetch in a tmux session named after the repo
name. using tmux for something like this frees up the calling process from
waiting for a long fetch, and also sets up the user to jump to the
session/more easily see what has been updated. kind of like using tmux
to grab the logs?